Refractor: `Participate` && Remove:`_getPositionValue` && Remove: `_getTokenUSDPrice`
The participants tokenAllocation based on their participation is now
calculated at the moment they participate, not after the ido finalizes.
logic has been removed as there is no exceed status anymore (see
previous commit).

`_getPositionValue` is now redundant, as it becomes a two liner,
calculated at participation and the value is stored in the Position
struct.

`-getTokenUSDPrice` removed, as no oracle is used anymore and we aim to
do go with 1 to 1. See previous removal of Oracle.